Moved some useful utilities out to top level utils
- Moved get_providers and some associated utilities to new top-level
  optimade.utils module, pre-empting a fledgling client utils module
- Added a get_child_databases_link() function
- Made assigning an ObjectID to a provider optional
